requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements.txt --no-hashes
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0 ; sys_platform != 'emscripten'
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
dotenv==0.9.9
    # via ai-automation-using-mcp
filetype==1.2.0
    # via langchain-google-genai
google-ai-generativelanguage==0.9.0
    # via langchain-google-genai
google-api-core==2.28.1
    # via google-ai-generativelanguage
google-auth==2.42.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grandalf==0.8
    # via ai-automation-using-mcp
greenlet==3.2.4
    # via playwright
grpcio==1.76.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain==1.0.2
    # via ai-automation-using-mcp
langchain-core==1.0.1
    # via
    #   langchain
    #   langchain-google-genai
    #   langchain-mcp-adapters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==3.0.0
    # via ai-automation-using-mcp
langchain-mcp-adapters==0.1.11
    # via ai-automation-using-mcp
langgraph==1.0.1
    # via langchain
langgraph-checkpoint==3.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.1
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.38
    # via langchain-core
mcp==1.19.0
    # via
    #   ai-automation-using-mcp
    #   langchain-mcp-adapters
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.11.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
playwright==1.55.0
    # via ai-automation-using-mcp
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==6.33.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.3
    # via
    #   ai-automation-using-mcp
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
    #   mcp
    #   pydantic-settings
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via mcp
pyee==13.0.0
    # via playwright
pyparsing==3.2.5
    # via grandalf
python-dotenv==1.2.1
    # via
    #   dotenv
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pywin32==311 ; sys_platform == 'win32'
    # via mcp
pyyaml==6.0.3
    # via langchain-core
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   google-api-core
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
sse-starlette==3.0.2
    # via mcp
starlette==0.49.1
    # via mcp
tenacity==9.1.2
    # via langchain-core
typing-extensions==4.15.0
    # via
    #   anyio
    #   grpcio
    #   langchain-core
    #   langchain-mcp-adapters
    #   pydantic
    #   pydantic-core
    #   pyee
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.38.0 ; sys_platform != 'emscripten'
    # via mcp
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith
